+++ selfdrive/test/process_replay/model_replay.py
@@ -9,7 +9,7 @@ import cereal.messaging as messaging
 from openpilot.common.params import Params
 from openpilot.system.hardware import PC
 from openpilot.selfdrive.manager.process_config import managed_processes
-from openpilot.selfdrive.test.openpilotci import BASE_URL, get_url
+from openpilot.tools.lib.openpilotci import BASE_URL, get_url
 from openpilot.selfdrive.test.process_replay.compare_logs import compare_logs, format_diff
 from openpilot.selfdrive.test.process_replay.process_replay import get_process_config, replay_process
 from openpilot.system.version import get_commit
@@ -143,7 +143,7 @@ if __name__ == "__main__":
     import requests
     import threading
     import http.server
-    from openpilot.selfdrive.test.openpilotci import upload_bytes
+    from openpilot.tools.lib.openpilotci import upload_bytes
     os.environ['MAPS_HOST'] = 'http://localhost:5000'
 
     class HTTPRequestHandler(http.server.BaseHTTPRequestHandler):
@@ -229,7 +229,7 @@ if __name__ == "__main__":
 
   # upload new refs
   if (update or failed) and not PC:
-    from openpilot.selfdrive.test.openpilotci import upload_file
+    from openpilot.tools.lib.openpilotci import upload_file
 
     print("Uploading new refs")
