+++ selfdrive/test/process_replay/test_debayer.py
@@ -8,7 +8,7 @@ import pyopencl as cl  # install with `PYOPENCL_CL_PRETEND_VERSION=2.0 pip insta
 
 from openpilot.system.hardware import PC, TICI
 from openpilot.common.basedir import BASEDIR
-from openpilot.selfdrive.test.openpilotci import BASE_URL
+from openpilot.tools.lib.openpilotci import BASE_URL
 from openpilot.system.version import get_commit
 from openpilot.system.camerad.snapshot.snapshot import yuv_to_rgb
 from openpilot.tools.lib.logreader import LogReader
@@ -171,7 +171,7 @@ if __name__ == "__main__":
 
   # upload new refs
   if update or (failed and TICI):
-    from openpilot.selfdrive.test.openpilotci import upload_file
+    from openpilot.tools.lib.openpilotci import upload_file
 
     print("Uploading new refs")
