+++ selfdrive/test/update_ci_routes.py
@@ -11,7 +11,7 @@ from tqdm import tqdm
 
 from openpilot.selfdrive.car.tests.routes import routes as test_car_models_routes
 from openpilot.selfdrive.test.process_replay.test_processes import source_segments as replay_segments
-from openpilot.selfdrive.test.openpilotci import (DATA_CI_ACCOUNT, DATA_CI_ACCOUNT_URL, OPENPILOT_CI_CONTAINER,
+from openpilot.tools.lib.openpilotci import (DATA_CI_ACCOUNT, DATA_CI_ACCOUNT_URL, OPENPILOT_CI_CONTAINER,
                                                   DATA_CI_CONTAINER, get_azure_credential, get_container_sas, upload_file)
 
 DATA_PROD_ACCOUNT = "commadata2"
