+++ tools/replay/unlog_ci_segment.py
@@ -12,7 +12,7 @@ from collections import defaultdict
 import cereal.messaging as messaging
 from openpilot.tools.lib.framereader import FrameReader
 from openpilot.tools.lib.logreader import LogReader
-from openpilot.selfdrive.test.openpilotci import get_url
+from openpilot.tools.lib.openpilotci import get_url
 
 IGNORE = ['initData', 'sentinel']
